removes duplicated code for examples and doc snippets
